live_caption_logger/requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0

